+++ map/filter-by.js
@@ -1,0 +1,8 @@
+
+import filter from './filter'
+
+
+export default function filter_by (bud)
+{
+	return filter(bud.sample)
+}