+++ test/index.test.js
@@ -6,6 +6,8 @@ describe('index', () =>
 {
 	it('exports', () =>
 	{
+		expect(Object.keys(index).length).eq(16)
+
 		expect(index.Bud).a('function')
 
 		expect(index.Nothing).a('symbol')
@@ -17,9 +19,11 @@ describe('index', () =>
 		expect(index.Fin).a('function')
 
 		expect(index.join).a('function')
+		expect(index.transfer).a('function')
 		expect(index.turnoff).a('function')
 		expect(index.resource).a('function')
 
+		expect(index.tap).a('function')
 		expect(index.asap).a('function')
 		expect(index.capture).a('function')
 		expect(index.concat).a('function')