+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `raf`.
 * `resource` is now in index.
 
 ## `0.5.0` — 2020-05-13