+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+## `0.6.0` — 2021-12-10
 * **Breaking change**: `filter_by` is moved to its own module `map/filter-by`.
 * Added `merge`.
 * Added `debug`.